Data simulations are now in DataGeneration.py
